// src/app/models/warehouses.ts
export class Warehouses {
    id: number;
    title: string;
    user_id: number;
    description: string;
    image: string;
    category_id: string;
    works_id: string;
    location_id: string;
    employee_id: string;
    is_featured: boolean;
    is_active: boolean;
    created_at: Date;
    date: string;
    files: string;

}